fixed pandapowerWrapper and little fixes for PyPSAWrapper
