+++ .gitignore
@@ -1,0 +1,5 @@
+results/*
+!results/.gitkeep
+__pycache__/
+target/
+docker-compose.yml
