+++ AvroSchemas/bus.avsc
@@ -7,7 +7,8 @@
       {"name": "p", "type": "string"},
       {"name": "q", "type": "string"},
       {"name": "v_mag_pu", "type": "string"},
-      {"name": "v_ang", "type": "string"}
+      {"name": "v_ang", "type": "string"},
+      {"name": "control", "type": "string"}
 
    ]
 }
