--- PandaPowerWrapper/config.properties
@@ -1,5 +1,0 @@
-[general]
-kafkaBroker=broker.kafka.svc.cluster.local:9092
-schemaRegistry=http://schema-registry.kafka.svc.cluster.local:8081
-resourceDir=/data/resources
-resultsDir=/data/results
