--- PandaPowerWrapper/requirements.txt
@@ -1,3 +1,0 @@
-pandapower>=2.11.0
-pandas
-numpy
