+++ _data/scenarios/scenario.json
@@ -1,0 +1,32 @@
+{
+    "scenarioID": "pandatest",
+    "domainReferences": {},
+    "simulationStart": 0,
+    "simulationEnd": 10000,
+    "execution": {
+        "randomSeed": 42,
+        "syncedParticipants": 1,
+        "constraints": "",
+        "priority": 0
+    },
+    "buildingBlocks": [
+        {
+            "instanceID": "pandapowerWrapper0",
+            "type": "pandapowerWrapper",
+            "layer": "micro",
+            "domain": "energy",
+            "stepLength": 1000,
+            "parameters": {},
+            "resources": {
+                "file:///examples/network.json": "Network"
+            },
+            "results": {},
+            "synchronized": true,
+            "isExternal": false,
+            "responsibilities": [],
+            "observers": []
+        }
+    ],
+    "translators": [],
+    "projectors": []
+}
